Delay start of django app for mongodb
When running in daemon mode, the django app tries to connect to mongodb
before mongo is actually up and running.  This causes the start script
to end with a non-0 code and the container stops running.  To fix this,
delay the start of the django app by some arbitrary time to allow
mongodb to fully startup.